Added 'cart total' component to react
