+++ client/components/index.js
@@ -7,4 +7,6 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as AllProducts} from './all-products'
-export {default as ViewCart} from './ViewCart'
+export {default as ViewCart} from './view-cart'
+export {default as CartTotal} from './cart-total'
+export {default as CartProduct} from './cart-product'
