+++ public/style.css
@@ -49,12 +49,34 @@ td {
   padding: 0.25rem 0rem;
 }
 
-.cart-button {
+.button {
   color: 000411;
-  background-color: #e1efe6;
   border-radius: 5%;
   box-shadow: 1px 1px 1px #000411;
   cursor: pointer;
   font-size: 0.9rem;
   padding: 0.4rem;
+}
+
+.remove {
+  background-color: #e1efe6;
+}
+
+.checkout {
+  margin-top: 0.5rem;
+  background-color: #160c28;
+  color: ivory;
+  position: relative;
+  left: 20%;
+}
+
+.bold {
+  font-weight: 600;
+}
+
+.total {
+  margin: 3rem auto 0rem auto;
+  padding: 0.75rem;
+  border: 2px #160c28 solid;
+  border-radius: 2%;
 }
